+++ CHANGELOG.md
@@ -2,11 +2,13 @@
 
 **Version 4.12**
 
+- Added better support for warnings in translated edits.
+- Updated validation engine version from 6.9 to 6.10.
 - Updated dependencies.
 
 **Version 4.11**
 
-- Updated validation engine version from 6.8 to 6.9 (should have been done when the metafile functions changed) 
+- Updated validation engine version from 6.8 to 6.9 (should have been done when the metafile functions changed). 
 - Updated dependencies.
 
 **Version 4.10**